Refactored DICOMPlugin.
